Flip display vertically.
On my display module, the image was mirror-reversed vertically so that (0,0) was at the bottom. Assuming that the SPI connector is at the top, that is, and the PCB silkscreen is the right way up. The image was being drawn at the bottom of the disply. This commit fixes that so the active portion of the display is at the top and the digits are readable.