Add petrol station style digits, in red.
Use a bitmap to draw digits in a style inspired by the price sign at a petrol station that I saw on Gloucester Road today. Draw in red, although some signs use green LEDs. Add #define for DIGIT_ROWS because the digits are not as tall as the display on this OLED, unlike the I2C OLED which is only 32 pixels in height.