Add RGB565 image and ']' command to draw it.
Same image as I've previously used with the PIC32 and an LCD, 64*64 pixels. Python script to convert the ASCII PPM file to a C header file. Same image file should be useable on the STM32F103 in the Blue Pill, so the image and the script are in the root of the repo.