Convert to using a colour framebuffer.
With the bigger RAM size on the STM32F411, we can afford a 32kB buffer for the full colour image. Render into the buffer in colour and update the OLED screen just with Y co-ord limits. Remove 'clrPixel()' and similar monochrome drawing functions. TODO: copy the RGB565 image into the buffer with a 'blit' operation.